Moved dropdownlist that was created in Tracking page to it own new page PackageChecking. PackageChecking allows the current Customers to pick between their Tracking Number IDs and view details on their packages. I deleted Customer_ShipmentDetail because it was an unneed page.
