+++ PostOffice_Application/Customer/Customer_Home.aspx.cs
@@ -16,61 +16,10 @@ namespace PostOffice_Application
             if (Session["CustomerID"] != null)
             {
 
-                try
-                {
-                    var cb = new SqlConnectionStringBuilder();
-                    cb.DataSource = "team-4-post-office-dbs.database.windows.net";
-                    cb.UserID = "ServerAdmin";
-                    cb.Password = "UmaThurman18";
-                    cb.InitialCatalog = "Post_Office";
-
-                    using (SqlConnection connection = new SqlConnection(cb.ConnectionString))
-                    {
-                        connection.Open();
-                        string query = "SELECT COUNT(1) FROM LOGIN L, USER_TYPE U WHERE L.Username=@Username AND L.Password=@Password AND U.User_Type_Name=@UserType AND L.User_Type_ID=U.User_Type_ID";
-                     
-
-                            string getid = "SELECT C.Customer_ID FROM CUSTOMER C, LOGIN L WHERE C.Email = L.Username";
-                            SqlCommand getcustid = new SqlCommand(getid, connection);
-                            SqlDataReader idreader = getcustid.ExecuteReader();
-                            Session["CustomerID"] = idreader.ToString();
-
-                            
-                        
-                        
-                    }
-                }
-                catch (SqlException ex)
-                {
-                    Console.WriteLine(ex.ToString());
-                }
+    
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-                WelcomeLabel.Text = "Welcome to our Post Office Customer HomePage! "+ Session["Username"].ToString() + " Also your customerID = " + Session["CustomerID"]; //Session[Username] give the username 
+                WelcomeLabel.Text = "Welcome to our Post Office Customer HomePage! "+ Session["Username"].ToString(); //Session[Username] give the username 
             }
             else
             {
