--- PostOffice_Application/Customer/Customer_ShipmentDetail.aspx.cs
@@ -1,24 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.UI;
-using System.Web.UI.WebControls;
-
-namespace PostOffice_Application.Customer
-{
-    public partial class Customer_ShipmentDetail : System.Web.UI.Page
-    {
-        protected void Page_Load(object sender, EventArgs e)
-        {
-
-        }
-
-        protected void btnLogoff_Click(object sender, EventArgs e)
-        {
-            Session.Abandon();
-            Session.Contents.RemoveAll();
-            Response.Redirect("../Start/Login.aspx");
-        }
-    }
-}
